rw queue,kernel thread ok

+++ netfilter_mod/wireway_node.h
@@ -48,7 +48,7 @@ typedef struct id_cache_control
 }id_cache_control;
 
 wireway_collector*  lookup_collector(char *name);
-int create_collector(char *name);
+unsigned long  create_collector(char *name);
 void *cache_id_lookup(id_cache_control *cache,unsigned long id);
 int cache_id_insert(id_cache_control *cache,unsigned long id,void *item);
 id_cache_control *create_id_cache(void);
